Add files via upload
So, the port scanner does SYN scanning, but it should be really straightforward to add different types (FIN, UDP, NULL, Xmas Tree, inverse...and if we really want to get after it, Dumb). It writes to a specified output file. Apologies if my code is in anyway difficult to read.